pyproject.toml
[tool.poetry]
name = "queue-send-email"
package-mode = false


[tool.poetry.dependencies]
python = "^3.11"
pip = "^25.0.1"
wheel = "^0.45.1"
setuptools = "^78.1.0"
azure-functions = "^1.21.3"
requests = "^2.32.3"
jinja2 = "^3.1.6"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.2"
pylint = "^3.3.6"
mypy = "^1.15.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
